requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
python-telegram-bot==21.6
Jinja2==3.1.4
python-multipart==0.0.9
apscheduler==3.11.1
